requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
cffi==1.15.0
cryptography==36.0.1
Django==3.0
django-cors-headers==3.10.1
django-filter==21.1
django-rest-knox==4.1.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
Pillow==8.4.0
pycparser==2.21
pytz==2021.3
sqlparse
tzdata==2021.5
djongo==1.3.6
pymongo==3.12.1
gunicorn==19.7.1
whitenoise==3.3.1
